BE-777 Fix an issue of authentication
Even if enable authentication, user can login by using any invalid
credential. Added some error handlings and e2e-test scenarios for
user authentication.

+++ app/platform/fabric/e2e-test/specs/stopexplorer.sh
@@ -4,11 +4,23 @@
 
 SCRIPTPATH="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 ROOTPATH="$( cd "$(dirname "$0")/../../../../.." >/dev/null 2>&1 ; pwd -P )"
+CLEANUP=1
+
+while getopts "k" opt; do
+  case "$opt" in
+  k)
+    echo "keep DB data and wallet"
+    CLEANUP=0
+    ;;
+  esac
+done
 
 pushd ${ROOTPATH}
 
 ./stop.sh
-docker-compose -f ./app/platform/fabric/e2e-test/docker-compose.yaml down -v
-rm -rf wallet/ logs/
+if [ $CLEANUP -eq 1 ]; then
+  docker-compose -f ./app/platform/fabric/e2e-test/docker-compose.yaml down -v
+  rm -rf wallet/ logs/
+fi
 
 popd
